% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAND.R
\name{HAND}
\alias{HAND}
\title{Fitting a Homophily Adjusted Network Disturbances Model}
\usage{
HAND(
  y,
  X,
  A,
  Usample,
  optim.control = list(maxit = 200),
  mu_b = rep(0, ncol(X)),
  s2_b = 2.25^2,
  mu_g = rep(0, dim(Usample)[3]),
  s2_g = 2.25^2,
  a_s2 = 2,
  b_s2 = 2,
  mu_rho = 0.36,
  sd_rho = 0.7,
  seed = 1,
  init_vec = NULL,
  CI_level = 0.95,
  verbose = F,
  Usample.maxiter = 100,
  Usample.eps = 1e-08
)
}
\arguments{
\item{y}{a numeric vector of responses.}

\item{X}{a matrix of covariates.}

\item{A}{a row-normalized adjacency matrix.}

\item{Usample}{a sample of latent locations to use as priors; an array of size (size, n, D), where size is the number of sample, n is the number of nodes in the network, and D is the number of latent dimension.}

\item{optim.control}{optional control parameters for \code{\link{optim}}.}

\item{mu_b}{a mean vector of the normal prior on \eqn{\beta}.}

\item{s2_b}{a numeric value for the variance of the normal prior on \eqn{\beta}.}

\item{mu_g}{a mean vector of the normal prior on \eqn{\gamma}.}

\item{s2_g}{a numeric value for the variance of the normal prior on \eqn{\gamma}.}

\item{a_s2}{two times the shape of the inverse-gamma prior on \ifelse{html}{\out{&sigma;<sup>2</sup>}}{\eqn{\sigma^2}}.}

\item{b_s2}{two times the scale of the inverse-gamma prior on \ifelse{html}{\out{&sigma;<sup>2</sup>}}{\eqn{\sigma^2}}.}

\item{mu_rho}{a mean vector of the truncated normal prior on \eqn{\rho}.}

\item{sd_rho}{a numeric value for the variance of the truncated normal prior on \eqn{\rho}.}

\item{seed}{random seed used in \code{\link{optim}}.}

\item{init_vec}{an initialization vector for \eqn{(\beta, \gamma,} \ifelse{html}{\out{&sigma;<sup>2</sup>}}{\eqn{\sigma^2}}, \eqn{\rho)} in \code{\link{optim}}. If not provided, the initial values are estimated using the two-stage least squares (TSLS) estimator proposed by Kelejian & Prucha (1998).}

\item{CI_level}{The posterior probability to be contained in the credible 
interval.}

\item{verbose}{a logical value indicating whether to print steps of the algorithm.}

\item{Usample.maxiter}{a integer value indicating the maximum iteration used in approximating Usample with a matrix normal distribution.}

\item{Usample.eps}{a numeric value indicating the precision used in approximating Usample with a matrix normal distribution.}
}
\description{
\code{HAND} implements a normal approximation to the posterior of the homophily
adjusted network disturbances (HAND) model.
}
\details{
\code{HAND} fits a homophily adjsuted network disturbances model given by
\ifelse{html}{\out{<p>y|A, X &#126; N (&Lambda;&gamma; + X&beta;,  &gamma;'&gamma; &Omega; + &sigma;<sup>2</sup> (I - &rho; A)<sup>-1</sup>(I - &rho; A')<sup>-1</sup>)</p>}}{\deqn{
  y|A, X ~ N (\Lambda\gamma + X\beta,  \gamma'\gamma \Omega + \sigma^2 (I - \rho A)^{-1}(I - \rho A')^{-1})
}}
where \eqn{\Lambda} and \eqn{\Omega} are MLE of the sample of U assuming a matrix normal distribution with mean \eqn{\Lambda} and row variance covariance \eqn{\Omega}.
}
\references{
Kelejian, H. H., & Prucha, I. R. (1998). A generalized spatial two-stage least squares procedure for estimating a spatial autoregressive model with autoregressive disturbances. The journal of real estate finance and economics, 17, 99-121.
}
\seealso{
\code{\link{optim}}
}
