% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEM.R
\name{NEM}
\alias{NEM}
\title{Fitting a network effects model}
\usage{
NEM(
  y,
  X,
  A,
  optim.control = list(maxit = 200),
  mu_b = rep(0, ncol(X)),
  s2_b = 2.25^2,
  a_s2 = 2,
  b_s2 = 2,
  mu_rho = 0.36,
  sd_rho = 0.7,
  seed = 1,
  init_vec = NULL,
  aa = 0.05
)
}
\arguments{
\item{y}{a numeric vector of responses.}

\item{X}{a matrix of covariates.}

\item{A}{a row-normalized adjacency matrix.}

\item{optim.control}{optional control parameters for \code{\link{optim}}.}

\item{mu_b}{a mean vector of the normal prior on \eqn{\beta}.}

\item{s2_b}{a numeric value for the variance of the normal prior on \eqn{\beta}.}

\item{a_s2}{two times the shape of the inverse-gamma prior on \ifelse{html}{\out{&sigma;<sup>2</sup>}}{\eqn{\sigma^2}}.}

\item{b_s2}{two times the scale of the inverse-gamma prior on \ifelse{html}{\out{&sigma;<sup>2</sup>}}{\eqn{\sigma^2}}.}

\item{mu_rho}{a mean vector of the truncated normal prior on \eqn{\rho}.}

\item{sd_rho}{a numeric value for the variance of the truncated normal prior on \eqn{\rho}.}

\item{seed}{random seed used in \code{\link{optim}}.}

\item{init_vec}{an initialization vector for \eqn{(\beta, \gamma,} \ifelse{html}{\out{&sigma;<sup>2</sup>}}{\eqn{\sigma^2}}, \eqn{\rho)} in \code{\link{optim}}. If not provided, the initial values are estimated using the two-stage least squares (TSLS) estimator proposed by Kelejian & Prucha (1998).}

\item{aa}{a numeric value corresponds to the significance level used to report the credible intervals of the parameter estimates.


#' @details
\code{NEM} fits a network effects model given by
\ifelse{html}{\out{<p>y|A, X &#126; N ( (I - &rho; A)<sup>-1</sup> X&beta;, &sigma;<sup>2</sup>(I - &rho; A)<sup>-1</sup>(I - &rho; A')<sup>-1</sup>)</p>}}{\deqn{
  y|A, X ~ N ( (I - \rho A)^{-1}X\beta,  \sigma^2 (I - \rho A)^{-1}(I - \rho A')^{-1})
}}}
}
\description{
\code{NEM} implements a normal approximation to the posterior of the network effects model.
}
\references{
Doreian, P. (1980). Linear models with spatially distributed data: Spatial disturbances or spatial effects?. Sociological Methods & Research, 9(1), 29-60.
Kelejian, H. H., & Prucha, I. R. (1998). A generalized spatial two-stage least squares procedure for estimating a spatial autoregressive model with autoregressive disturbances. The journal of real estate finance and economics, 17, 99-121.
}
\seealso{
\code{\link{optim}}
}
