% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.HANE}
\alias{print.HAND}
\title{Print hanam objects.}
\usage{
\method{print}{HANE}(x)

\method{print}{HAND}(x)
}
\arguments{
\item{x}{an object used to select a method.}
}
\description{
Print hanam objects.
}
