% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.HANE}
\alias{summary.HAND}
\title{Summary functions for hanam objects}
\usage{
\method{summary}{HANE}(object, CI_level = 0.95)

\method{summary}{HAND}(object, CI_level = 0.95)
}
\arguments{
\item{object}{HANE or HAND object}

\item{CI_level}{Posterior probability covered by credible interval}
}
\description{
Summary functions for hanam objects
}
